refactor: extract Recovery tab to views/RecoveryView.jsx + test (PR6)
Second view extraction of the week-2 burndown (#52). Moves the `recovery`
tab — readiness composite, HRV/RHR/sleep trend charts, blood-pressure log,
niggle/injury watch, and the lipid/hormone reference panels — out of the
erg-dashboard.jsx monolith into a self-contained component.

- web/src/views/RecoveryView.jsx: the recovery JSX, verbatim. Takes the live
  training-load snapshot (`latest`, only latest.tsb is read) and `isWide`
  (viewport flag) as props; imports its static data + chart deps directly.
- web/src/constants/logs.js: LIPID_REF, HORMONE_REF, NIGGLES (recovery-only
  data, were inline in the monolith) moved here with the other recovery
  tables. They are recovery-only, so they leave the monolith entirely.
- web/src/erg-dashboard.jsx: the ~780-line block becomes
  `{view === 'recovery' && <RecoveryView latest={latest} isWide={isWide} />}`;
  the now-unused bpCategory/bpLog/bloodsLog imports are pruned.

Adds web/src/views/__tests__/RecoveryView.test.jsx (9 tests). Mocks
recoveryLog to drive the readiness branches (empty state + READY/CAUTION/REST
+ missing sleep score + positive-TSB narrative) and asserts the section
headers, lipid/hormone panels, and niggle rows render. Coverage: 100% lines /
100% funcs / 62.5% branch (in line with the other large extracted views).

Build, full Vitest suite (290 passing), lint, and format:check all green.

Closes #73
